Source/requirements.txt
Flask
python-dotenv
requests
# mypy